Refactor test cases
All tests are moved to basic-worker-actions.bal. This is because all
the tests are just testsing types.
