+++ tests/jballerina-unit-test/src/test/resources/test-src/workers/sync-send.bal
@@ -1,5 +1,4 @@
 import ballerina/jballerina.java;
-import ballerina/lang.'value as value;
 
 string append = "";
 function simpleSyncSend() returns string {
@@ -629,211 +628,6 @@ public function testFailureForReceiveWithError() returns boolean {
     return x.w1 is E2 && x.w2 === true;
 }
 
-string appendCloneable = "";
-function testSimpleSyncSendWithCloneableExpression() {
-    string done = processWithCloneableExpression();
-    assertValueEquality("done", done);
-    assertValueEquality("w2cw2cw2cw2cw2cw1cw1cw1cw1cw1c", appendCloneable);
-}
-
-function processWithCloneableExpression() returns string {
-    worker w1Cloneable returns boolean {
-        value:Cloneable w1a = 10;
-        value:Cloneable w1b = 11;
-        value:Cloneable w1c = 12;
-        w1a -> w2Cloneable;
-        w1b ->> w2Cloneable;
-        w1c -> w2Cloneable;
-        foreach var i in 1 ... 5 {
-            appendCloneable = appendCloneable + "w1c";
-        }
-        return true;
-    }
-
-    worker w2Cloneable returns boolean {
-        value:Cloneable w2a = 15;
-        value:Cloneable w2b = 15;
-        value:Cloneable w2c = 15;
-        sleep(10);
-        foreach var i in 1 ... 5 {
-            appendCloneable = appendCloneable + "w2c";
-        }
-        w2a = <- w1Cloneable;
-        w2b = <- w1Cloneable;
-        w2c = <- w1Cloneable;
-        if (w2a is int) {
-            assertValueEquality(10, w2a);
-        } else {
-            return false;
-        }
-        if (w2b is int) {
-            assertValueEquality(11, w2b);
-        } else {
-            return false;
-        }
-        if (w2c is int) {
-            assertValueEquality(12, w2c);
-        } else {
-            return false;
-        }
-        return true;
-    }
-
-    record { boolean w1Cloneable; boolean w2Cloneable; } x = wait {w1Cloneable, w2Cloneable};
-    assertValueEquality(true, x.w1Cloneable);
-    assertValueEquality(true, x.w2Cloneable);
-    return "done";
-}
-
-function simpleSyncSendErrorType() returns string {
-    @strand {thread: "parent"}
-    worker w1Error returns error {
-        error w1a = error("10");
-        error w1b = error("11");
-        error w1c = error("12");
-        w1a -> w2Error;
-        w1b ->> w2Error;
-        w1c -> w2Error;
-        foreach var i in 1 ... 5 {
-            appendErrorMessages = appendErrorMessages + "w1e";
-        }
-        return error("Completed");
-    }
-
-    @strand {thread: "parent"}
-    worker w2Error returns error {
-        error w2a = error("15");
-        error w2b = error("15");
-        error w2c = error("15");
-        sleep(10);
-        foreach var i in 1 ... 5 {
-            appendErrorMessages = appendErrorMessages + "w2e";
-        }
-        w2a = <- w1Error;
-        w2b = <- w1Error;
-        w2c = <- w1Error;
-        assertValueEquality("10", w2a.message());
-        assertValueEquality("11", w2b.message());
-        assertValueEquality("12", w2c.message());
-        return error("Completed");
-    }
-
-    record {
-        error w1Error;
-        error w2Error;
-    } x = wait {w1Error, w2Error};
-    assertValueEquality("Completed", x.w1Error.message());
-    assertValueEquality("Completed", x.w2Error.message());
-    return "done";
-}
-
-function simpleSyncSendXMLType() returns string {
-    @strand {thread: "parent"}
-    worker w1Xml returns xml {
-        xml w1a = xml `10`;
-        xml w1b = xml `11`;
-        xml w1c = xml `12`;
-        w1a -> w2Xml;
-        w1b ->> w2Xml;
-        w1c -> w2Xml;
-        foreach var i in 1 ... 2 {
-            xmlValue = xmlValue.concat(w1a);
-        }
-        return xml `Completed`;
-    }
-
-    @strand {thread: "parent"}
-    worker w2Xml returns xml {
-        xml w2a = xml `boom`;
-        xml w2b = xml `boom`;
-        xml w2c = xml `boom`;
-        sleep(10);
-        foreach var i in 1 ... 2 {
-            xmlValue = xmlValue.concat(w2a);
-        }
-        w2a = <- w1Xml;
-        w2b = <- w1Xml;
-        w2c = <- w1Xml;
-        assertValueEquality(xml `10`, w2a);
-        assertValueEquality(xml `11`, w2b);
-        assertValueEquality(xml `12`, w2c);
-        return xml `Completed`;
-    }
-
-    record {
-        xml w1Xml;
-        xml w2Xml;
-    } x = wait {w1Xml, w2Xml};
-    assertValueEquality("Completed", x.w1Xml.toString());
-    assertValueEquality("Completed", x.w2Xml.toString());
-    return "done";
-}
-
-type IntRec readonly & record {
-                           int ID = 0;
-                       };
-
-string readonlyRecIDs = "START";
-
-function testSimpleSyncSendWithReadonlyRecord() {
-    assertValueEquality("done", simpleSyncSendReadonllyRecord());
-    assertValueEquality("START101033", readonlyRecIDs.toString());
-}
-
-xml xmlValue = xml ``;
-function testSimpleSyncSendWithXMLType() {
-    assertValueEquality("done", simpleSyncSendXMLType());
-    assertValueEquality("boomboom1010", xmlValue.toString());
-}
-
-string appendErrorMessages = "";
-function testSimpleSyncSendWithErrorType() {
-    assertValueEquality("done", simpleSyncSendErrorType());
-    assertValueEquality("w2ew2ew2ew2ew2ew1ew1ew1ew1ew1e", appendErrorMessages);
-}
-
-function simpleSyncSendReadonllyRecord() returns string {
-    @strand {thread: "parent"}
-    worker w1readonlyRec returns string {
-        IntRec w1a = {ID: 10};
-        IntRec w1b = {ID: 11};
-        IntRec w1c = {ID: 12};
-        foreach var i in 1 ... 2 {
-            readonlyRecIDs = readonlyRecIDs.concat(w1a.ID.toString());
-        }
-        w1a -> w2readonlyRec;
-        w1b ->> w2readonlyRec;
-        w1c -> w2readonlyRec;
-        return "Completed";
-    }
-
-    @strand {thread: "parent"}
-    worker w2readonlyRec returns string {
-        IntRec w2a = {ID: 3};
-        IntRec w2b;
-        IntRec w2c;
-        sleep(10);
-        foreach var i in 1 ... 2 {
-            readonlyRecIDs = readonlyRecIDs.concat(w2a.ID.toString());
-        }
-        w2a = <- w1readonlyRec;
-        w2b = <- w1readonlyRec;
-        w2c = <- w1readonlyRec;
-        assertValueEquality(10, w2a.ID);
-        assertValueEquality(11, w2b.ID);
-        assertValueEquality(12, w2c.ID);
-        return "Completed";
-    }
-
-    record {
-        string w1readonlyRec;
-        string w2readonlyRec;
-    } x = wait {w1readonlyRec, w2readonlyRec};
-    assertValueEquality("Completed", x.w1readonlyRec);
-    assertValueEquality("Completed", x.w2readonlyRec);
-    return "done";
-}
-
 function getFalse() returns boolean {
     return false;
 }
@@ -854,15 +648,4 @@ function validateError(any|error value, string message) {
         panic error("Expected error message: " + message + ", found: " + value.message());
     }
     panic error("Expected error, found: " + (typeof value).toString());
-}
-
-type AssertionError distinct error;
-const ASSERTION_ERROR_REASON = "AssertionError";
-
-function assertValueEquality(anydata expected, anydata actual) {
-    if expected == actual {
-        return;
-    }
-    panic error(ASSERTION_ERROR_REASON,
-                message = "expected '" + expected.toString() + "', found '" + actual.toString () + "'");
 }
